General Gem Cleanup
 - Renamed log_out_request to logout_request.
 - Upgraded rspec.
 - Made some general gemspec changes.
 - Removed debugger gems.
