+++ Gemfile
@@ -2,9 +2,4 @@ source 'https://rubygems.org/'
 
 gemspec
 
-group :test do
-  gem 'rake'
-  gem 'ruby-debug', :platform => :ruby_18
-  gem 'debugger', :platform => :ruby_19
-  gem 'rspec', '1.3.2'
-end
+
