+++ lib/onelogin/saml.rb
@@ -42,5 +42,6 @@ require 'onelogin/saml/settings'
 require 'onelogin/saml/name_identifiers'
 require 'onelogin/saml/status_codes'
 require 'onelogin/saml/meta_data'
+require 'onelogin/saml/logout_request'
 require 'onelogin/saml/log_out_request'
 require 'onelogin/saml/logout_response'
