+++ ruby-saml-mod.gemspec
@@ -1,33 +1,21 @@
+# coding: utf-8
+
 Gem::Specification.new do |s|
   s.name = %q{ruby-saml-mod}
-  s.version = "0.1.30"
-
+  s.version = "0.1.31"
   s.authors = ["OneLogin LLC", "Bracken", "Zach", "Cody", "Jeremy", "Paul", "Nick"]
-  s.date = %q{2014-05-05}
-  s.extra_rdoc_files = [
-    "LICENSE"
-  ]
-  s.files = [
-    "LICENSE",
-    "README",
-    "lib/onelogin/saml.rb",
-    "lib/onelogin/saml/auth_request.rb",
-    "lib/onelogin/saml/authn_contexts.rb",
-    "lib/onelogin/saml/log_out_request.rb",
-    "lib/onelogin/saml/logout_response.rb",
-    "lib/onelogin/saml/meta_data.rb",
-    "lib/onelogin/saml/name_identifiers.rb",
-    "lib/onelogin/saml/response.rb",
-    "lib/onelogin/saml/settings.rb",
-    "lib/onelogin/saml/status_codes.rb",
-    "lib/xml_sec.rb",
-    "ruby-saml-mod.gemspec"
-  ]
+  s.summary = %q{Ruby library for SAML service providers}
+  s.homepage = %q{http://github.com/instructure/ruby-saml}
+  s.description = %q{This is an early fork from https://github.com/onelogin/ruby-saml - I plan to "rebase" these changes ontop of their current version eventually. }
+  s.date = Time.now.strftime("%Y-%m-%d")
+
+  s.files = Dir.glob("{lib,spec}/**/*")
+  s.test_files = s.files.grep(%r{^(test|spec|features)/})
+  s.require_paths = ["lib"]
+
   s.add_dependency('libxml-ruby', '>= 2.3.0')
   s.add_dependency('ffi')
 
-  s.homepage = %q{http://github.com/bracken/ruby-saml}
-  s.require_paths = ["lib"]
-  s.summary = %q{Ruby library for SAML service providers}
-  s.description = %q{This is an early fork from https://github.com/onelogin/ruby-saml - I plan to "rebase" these changes ontop of their current version eventually. }
+  s.add_development_dependency 'rake'
+  s.add_development_dependency 'rspec', '2.14.1'
 end
