+++ spec/logout_response_spec.rb
@@ -1,5 +1,4 @@
-require File.expand_path(File.dirname(__FILE__) + '/spec_helper.rb')
-
+require 'spec_helper'
 require 'rexml/document'
 require 'cgi'
